+++ pkg/webhook/constants.go
@@ -18,6 +18,8 @@ const (
 	//--- Vault Sidecar Injector annotation keys (without prefix)
 	vaultInjectorAnnotationInjectKey          = "inject"              // Mandatory
 	vaultInjectorAnnotationAuthMethodKey      = "auth"                // Optional. Vault Auth Method to use: kubernetes (default) or approle
+	vaultInjectorAnnotationModeKey            = "mode"                // Optional. Comma-separated list of mode(s) to enable.
+	vaultInjectorAnnotationProxyPortKey       = "proxy-port"          // Optional. Port assigned to local Vault proxy.
 	vaultInjectorAnnotationRoleKey            = "role"                // Optional. To explicitly provide Vault role to use
 	vaultInjectorAnnotationSATokenKey         = "sa-token"            // Optional. Full path to service account token used for Vault Kubernetes authentication
 	vaultInjectorAnnotationSecretsPathKey     = "secrets-path"        // Optional. Full path, e.g.: "secret/<some value>", "aws/creds/<some role>", ... Several values separated by ','.
@@ -32,6 +34,10 @@ const (
 	vaultInjectorAnnotationStatusValue      = "injected"
 	vaultInjectorAnnotationWorkloadJobValue = "job"
 
+	//--- Vault Sidecar Injector supported modes
+	vaultInjectorModeSecrets = "secrets" // Enable dynamic fetching of secrets from Vault store
+	vaultInjectorModeProxy   = "proxy"   // Enable local Vault proxy
+
 	//--- Vault Sidecar Injector mount path for service accounts
 	vaultInjectorSATokenVolMountPath = "/var/run/secrets/talend/vault-sidecar-injector/serviceaccount"
 	k8sDefaultSATokenVolMountPath    = "/var/run/secrets/kubernetes.io/serviceaccount"
@@ -39,6 +45,8 @@ const (
 	//--- Vault Agent placeholders
 	vaultRolePlaceholder                 = "<APP_VAULT_ROLE>"
 	vaultAuthMethodPlaceholder           = "<APPSVC_VAULT_AUTH_METHOD>"
+	vaultProxyConfigPlaceholder          = "<APPSVC_PROXY_CONFIG>"
+	vaultProxyPortPlaceholder            = "<APPSVC_PROXY_PORT>"
 	vaultAppSvcSecretsPathPlaceholder    = "<APPSVC_VAULT_SECRETS_PATH>"
 	templateAppSvcDestinationPlaceholder = "<APPSVC_SECRETS_DESTINATION>"
 	templateContentPlaceholder           = "<APPSVC_TEMPLATE_CONTENT>"
@@ -46,11 +54,12 @@ const (
 	templateTemplatesPlaceholder         = "<APPSVC_TEMPLATES>"
 	appSvcSecretsVolMountPathPlaceholder = "<APPSVC_SECRETS_VOL_MOUNTPATH>"
 
-	vaultDefaultAuthMethod            = "kubernetes"         // Default auth method used by Vault Agent
+	vaultK8sAuthMethod                = "kubernetes"         // Default auth method used by Vault Agent
 	appSvcSecretsVolName              = "secrets"            // Name of the volume shared between containers to store secrets file(s)
 	templateAppSvcDefaultDestination  = "secrets.properties" // Default secrets destination
 	k8sDefaultServiceAccountTokenPath = "/var/run/secrets/kubernetes.io/serviceaccount/token"
 	vaultDefaultSecretsEnginePath     = "secret" // Default path for Vault K/V Secrets Engine if no 'secrets-path' annotation
+	vaultProxyDefaultPort             = "8200"   // Default port to access local Vault proxy
 
 	//--- Job handling - Temporary mechanism until KEP https://github.com/kubernetes/enhancements/blob/master/keps/sig-apps/sidecarcontainers.md is implemented (and we migrate on appropriate version of k8s)
 	jobMonitoringContainerName     = "tvsi-job-babysitter" // Name of our specific sidecar container to inject in submitted jobs